requirements.txt
streamlit==1.44.1
tensorflow-cpu==2.19.0  
matplotlib==3.10.1
numpy==2.2.4 
pandas==2.2.3
streamlit_drawable_canvas==0.9.3
pytest==8.3.5
coverage==7.8.0